LayoutTests/media/modern-media-controls/status-support/status-support-playing-expected.txt

Testing the StatusSupport behavior when media plays.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.statusLabel.text is ""

PASS successfullyParsed is true

TEST COMPLETE
